:wrench:chore: separate essential form to atomic
